+++ src/buffer.h
@@ -6,8 +6,8 @@
 typedef struct _buffer {
   char *buffer;
   char static_buffer[1024];
-  int pos;
-  int buffsize;
+  size_t pos;
+  size_t buffsize;
 } BUFFER;
 
 #define buffer_init(buff) ((buff)->buffer = (char*)&(buff)->static_buffer, (buff)->pos = 0, (buff)->buffsize = 1024)
@@ -20,17 +20,17 @@ typedef struct _buffer {
 #define buffer_reset(buff) ((buff)->pos = 0)
 
 
-void buffer_openspacehelper(BUFFER *buff, int amount);
-void buffer_resize(BUFFER* buff, int size);
+void buffer_openspacehelper(BUFFER *buff, size_t amount);
+void buffer_resize(BUFFER* buff, size_t size);
 
 #define buffer_free(buff)	buffer_resize(buff, 0)
 
 #define buffer_openspace(buff, amount) \
-  if (((int)(amount) + (buff)->pos) > (buff)->buffsize) \
+  if (((size_t)(amount) + (buff)->pos) > (buff)->buffsize) \
     buffer_openspacehelper((buff), (amount));
 
 #define buffer_addchar(buff, c) { buffer_openspace(buff, 1); (buff)->buffer[(buff)->pos] = (c); (buff)->pos++; }
-#define buffer_addstring(buff, str, len) { buffer_openspace((buff), (len)); memcpy((buff)->buffer + (buff)->pos, str, (len)); (buff)->pos += (int)(len); }
+#define buffer_addstring(buff, str, len) { buffer_openspace((buff), (len)); memcpy((buff)->buffer + (buff)->pos, str, (len)); (buff)->pos += (size_t)(len); }
 #define buffer_putchar(buff, c) (buff)->buffer[(buff)->pos] = (c)
 #define buffer_putstring(buff, str, len) { buffer_openspace((buff), (len)); memcpy((buff)->buffer + (buff)->pos, str, (len)); }
 #define buffer_setpos(buff, newpos) (buff)->pos = newpos;
